Several fixes:
- converted to Maven project
- introduced log4j
- refactored CpnToUnit.java
- refactored classes in package cpn
