+++ src/main/java/hlomozda/cpnunittransformer/definitions/GlobalXmlDefinitions.java
