+++ src/main/java/hlomozda/cpnunittransformer/definitions/TransitionXmlDefinitions.java
