+++ src/main/java/hlomozda/cpnunittransformer/geom/Rectangle.java
