+++ src/main/java/hlomozda/cpnunittransformer/geom/Segment.java
