+++ src/main/java/hlomozda/cpnunittransformer/tss/CpnShldiSolver.java
