+++ src/main/java/hlomozda/cpnunittransformer/tss/CpnShldiSolverRunner.java
