+++ src/main/java/hlomozda/cpnunittransformer/tss/CpnTssFinderRunner.java
