+++ src/main/java/hlomozda/cpnunittransformer/utils/IncidenceMatrixBuilder.java
