+++ src/main/java/hlomozda/cpnunittransformer/utils/VectorSpaceUtils.java
